+++ src/Graphics/Vty/Platform/Windows.hs
@@ -38,5 +38,5 @@ mkVtyWithSettings userConfig settings = do
                                 (configTermWidthMaps userConfig)
 
     input <- buildInput userConfig settings
-    out <- buildOutput settings
+    out <- buildOutput userConfig settings
     mkVtyFromPair input out